refactor: renames and moving code
