+++ grpc/client_test.go
@@ -1,4 +1,4 @@
-package common
+package grpc
 
 import (
 	"errors"
